+++ src/main/java/com/sedikev/domain/entity/CarteraEntity.java
@@ -24,5 +24,5 @@ public class CarteraEntity {
     private UsuarioEntity usuarioEntity;
 
     @Column(name = "saldo")
-    private BigDecimal saldo;
+    private Float saldo;
 }
